Refactor TranscriptionPanel into hooks
