+++ src/components/consultation/TranscriptionPanel.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect } from 'react';
+import React from 'react';
 import { motion } from 'framer-motion';
 import {
   StopIcon,
@@ -7,7 +7,7 @@ import {
   CheckCircleIcon,
   SparklesIcon
 } from '@heroicons/react/24/outline';
-import { useConsultation } from '../../contexts/ConsultationContext';
+import { useTranscription } from '../../../hooks/useTranscription';
 
 const TranscriptionPanel = () => {
   const {
@@ -16,427 +16,18 @@ const TranscriptionPanel = () => {
     liveTranscript,
     finalTranscript,
     confidence,
-    startRecording,
-    stopRecording,
-    updateLiveTranscript,
-    finalizeTranscript,
-    logEvent,
-    addAiMessage,
-    isAdvancedMode
-  } = useConsultation();
+    recordingTime,
+    audioLevel,
+    transcriptionService,
+    micPermission,
+    micDiagnostics,
+    handleStartRecording,
+    handleStopRecording,
+    checkMicrophonePermission,
+    runMicrophoneDiagnostics,
+    isAdvancedMode,
+  } = useTranscription();
   
-  const [recordingTime, setRecordingTime] = useState(0);
-  const [micPermission, setMicPermission] = useState('prompt'); // 'granted' | 'denied' | 'prompt' | 'checking'
-  const [audioLevel, setAudioLevel] = useState(0);
-  const [transcriptionService] = useState('browser'); // 'browser' | 'whisper'
-  const [micDiagnostics, setMicDiagnostics] = useState(null);
-  
-  const mediaRecorderRef = useRef(null);
-  const audioChunksRef = useRef([]);
-  const intervalRef = useRef(null);
-  const audioContextRef = useRef(null);
-  const analyserRef = useRef(null);
-  const recognitionRef = useRef(null);
-  
-  // Check microphone permission on mount
-  useEffect(() => {
-    checkMicrophonePermission();
-    return () => {
-      if (intervalRef.current) clearInterval(intervalRef.current);
-      if (recognitionRef.current) recognitionRef.current.stop();
-    };
-  }, []);
-  
-  // Update recording time
-  useEffect(() => {
-    if (isRecording && !isPaused) {
-      intervalRef.current = setInterval(() => {
-        setRecordingTime(prev => prev + 1);
-      }, 1000);
-    } else {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-      }
-    }
-    
-    return () => {
-      if (intervalRef.current) clearInterval(intervalRef.current);
-    };
-  }, [isRecording, isPaused]);
-  
-  const runMicrophoneDiagnostics = async () => {
-    const diagnostics = {
-      timestamp: new Date().toISOString(),
-      browser: navigator.userAgent,
-      isHTTPS: location.protocol === 'https:',
-      hasMediaDevices: !!navigator.mediaDevices,
-      hasGetUserMedia: !!navigator.mediaDevices?.getUserMedia,
-      hasPermissionsAPI: !!navigator.permissions,
-      devices: [],
-      permissionState: 'unknown',
-      errors: [],
-      recommendations: []
-    };
-
-    try {
-      // Check for basic API availability
-      if (!navigator.mediaDevices) {
-        diagnostics.errors.push('MediaDevices API no disponible en este navegador');
-        diagnostics.recommendations.push('Actualiza tu navegador a una versión más reciente');
-      }
-
-      if (!navigator.mediaDevices?.getUserMedia) {
-        diagnostics.errors.push('getUserMedia API no disponible');
-        diagnostics.recommendations.push('Tu navegador no soporta acceso al micrófono');
-      }
-
-      // Check HTTPS requirement
-      if (!diagnostics.isHTTPS && location.hostname !== 'localhost') {
-        diagnostics.errors.push('Se requiere HTTPS para acceso al micrófono');
-        diagnostics.recommendations.push('Accede al sitio usando HTTPS://');
-      }
-
-      // Enumerate audio devices
-      try {
-        const devices = await navigator.mediaDevices.enumerateDevices();
-        const audioInputs = devices.filter(device => device.kind === 'audioinput');
-        diagnostics.devices = audioInputs.map(device => ({
-          deviceId: device.deviceId,
-          label: device.label || 'Micrófono sin nombre',
-          groupId: device.groupId
-        }));
-        
-        if (audioInputs.length === 0) {
-          diagnostics.errors.push('No se encontraron dispositivos de audio de entrada');
-          diagnostics.recommendations.push('Conecta un micrófono o verifica que esté habilitado');
-        }
-        
-        // Store devices in diagnostics (setAvailableDevices removed to fix unused variable warning)
-      } catch (deviceError) {
-        diagnostics.errors.push(`Error enumerando dispositivos: ${deviceError.message}`);
-      }
-
-      // Check permissions
-      if (navigator.permissions) {
-        try {
-          const permissionResult = await navigator.permissions.query({ name: 'microphone' });
-          diagnostics.permissionState = permissionResult.state;
-          
-          switch (permissionResult.state) {
-            case 'denied':
-              diagnostics.errors.push('Permisos de micrófono denegados explícitamente');
-              diagnostics.recommendations.push('Haz clic en el ícono de candado en la barra de direcciones y permite el micrófono');
-              break;
-            case 'prompt':
-              diagnostics.recommendations.push('Los permisos se solicitarán al intentar grabar');
-              break;
-            case 'granted':
-              diagnostics.recommendations.push('Permisos concedidos, el micrófono debería funcionar');
-              break;
-          }
-        } catch (permError) {
-          diagnostics.errors.push(`Error verificando permisos: ${permError.message}`);
-        }
-      }
-
-      // Test actual microphone access
-      try {
-        const stream = await navigator.mediaDevices.getUserMedia({ 
-          audio: {
-            echoCancellation: true,
-            noiseSuppression: true,
-            autoGainControl: true
-          } 
-        });
-        
-        diagnostics.accessTest = 'success';
-        diagnostics.permissionState = 'granted';
-        setMicPermission('granted');
-        
-        // Test audio context
-        try {
-          const audioContext = new (window.AudioContext || window.webkitAudioContext)();
-          diagnostics.audioContextSupport = audioContext.state;
-          audioContext.close();
-        } catch (audioError) {
-          diagnostics.errors.push(`Audio Context no disponible: ${audioError.message}`);
-        }
-        
-        // Stop test stream
-        stream.getTracks().forEach(track => track.stop());
-        
-      } catch (accessError) {
-        diagnostics.accessTest = 'failed';
-        diagnostics.accessError = {
-          name: accessError.name,
-          message: accessError.message
-        };
-
-        switch (accessError.name) {
-          case 'NotAllowedError':
-            setMicPermission('denied');
-            diagnostics.errors.push('Acceso al micrófono denegado por el usuario');
-            diagnostics.recommendations.push('Permite el acceso al micrófono en el navegador');
-            diagnostics.recommendations.push('Revisa la configuración de permisos del sitio');
-            break;
-          case 'NotFoundError':
-            setMicPermission('denied');
-            diagnostics.errors.push('No se encontró ningún dispositivo de micrófono');
-            diagnostics.recommendations.push('Verifica que el micrófono esté conectado y funcionando');
-            diagnostics.recommendations.push('Revisa la configuración de audio del sistema');
-            break;
-          case 'NotReadableError':
-            diagnostics.errors.push('Micrófono en uso por otra aplicación');
-            diagnostics.recommendations.push('Cierra otras aplicaciones que puedan estar usando el micrófono');
-            diagnostics.recommendations.push('Reinicia el navegador e intenta nuevamente');
-            break;
-          case 'OverconstrainedError':
-            diagnostics.errors.push('Configuración de audio no soportada');
-            diagnostics.recommendations.push('El micrófono no soporta la configuración requerida');
-            break;
-          case 'SecurityError':
-            diagnostics.errors.push('Error de seguridad al acceder al micrófono');
-            diagnostics.recommendations.push('Verifica que el sitio tenga permisos de micrófono');
-            break;
-          default:
-            diagnostics.errors.push(`Error desconocido: ${accessError.message}`);
-            diagnostics.recommendations.push('Intenta recargar la página o reiniciar el navegador');
-        }
-      }
-
-    } catch (error) {
-      diagnostics.errors.push(`Error general en diagnóstico: ${error.message}`);
-    }
-
-    setMicDiagnostics(diagnostics);
-    return diagnostics;
-  };
-
-  const checkMicrophonePermission = async () => {
-    try {
-      // Only check permission status without requesting access
-      if ('permissions' in navigator) {
-        const result = await navigator.permissions.query({ name: 'microphone' });
-        setMicPermission(result.state);
-        
-        // Listen for permission changes
-        result.addEventListener('change', () => {
-          setMicPermission(result.state);
-        });
-      } else {
-        // If permissions API not available, assume we need to prompt
-        setMicPermission('prompt');
-      }
-    } catch (error) {
-      console.log('Permission API not available, will check on first recording attempt');
-      // Don't assume denied, just keep as prompt
-      setMicPermission('prompt');
-    }
-  };
-  
-  const setupAudioAnalysis = (stream) => {
-    audioContextRef.current = new (window.AudioContext || window.webkitAudioContext)();
-    analyserRef.current = audioContextRef.current.createAnalyser();
-    const microphone = audioContextRef.current.createMediaStreamSource(stream);
-    
-    analyserRef.current.fftSize = 256;
-    const bufferLength = analyserRef.current.frequencyBinCount;
-    const dataArray = new Uint8Array(bufferLength);
-    
-    microphone.connect(analyserRef.current);
-    
-    const updateAudioLevel = () => {
-      if (analyserRef.current) {
-        analyserRef.current.getByteFrequencyData(dataArray);
-        const average = dataArray.reduce((a, b) => a + b) / bufferLength;
-        setAudioLevel(average);
-        
-        if (isRecording) {
-          requestAnimationFrame(updateAudioLevel);
-        }
-      }
-    };
-    
-    updateAudioLevel();
-  };
-  
-  const setupBrowserTranscription = () => {
-    if (!('webkitSpeechRecognition' in window) && !('SpeechRecognition' in window)) {
-      console.warn('Speech recognition not supported');
-      return;
-    }
-    
-    const SpeechRecognition = window.SpeechRecognition || window.webkitSpeechRecognition;
-    recognitionRef.current = new SpeechRecognition();
-    
-    recognitionRef.current.continuous = true;
-    recognitionRef.current.interimResults = true;
-    recognitionRef.current.lang = 'es-ES';
-    
-    recognitionRef.current.onresult = (event) => {
-      let interimTranscript = '';
-      let finalTranscriptUpdate = '';
-      
-      for (let i = event.resultIndex; i < event.results.length; i++) {
-        const transcript = event.results[i][0].transcript;
-        const confidence = event.results[i][0].confidence || 0.8;
-        
-        if (event.results[i].isFinal) {
-          finalTranscriptUpdate += transcript;
-        } else {
-          interimTranscript += transcript;
-        }
-        
-        // Use confidence for quality indication
-        if (confidence < 0.5) {
-          console.warn('Low confidence transcription:', transcript, confidence);
-        }
-      }
-      
-      if (interimTranscript) {
-        updateLiveTranscript(interimTranscript, confidence);
-      }
-      
-      if (finalTranscriptUpdate) {
-        finalizeTranscript(finalTranscriptUpdate);
-        
-        // Send to AI if in advanced mode
-        if (isAdvancedMode && finalTranscriptUpdate.trim().length > 10) {
-          sendToAI(finalTranscriptUpdate);
-        }
-      }
-    };
-    
-    recognitionRef.current.onerror = (event) => {
-      console.error('Speech recognition error:', event.error);
-      logEvent('transcription_error', { error: event.error });
-    };
-    
-    recognitionRef.current.start();
-  };
-  
-  const sendToAI = async (transcript) => {
-    if (!isAdvancedMode) return;
-    
-    // Add context message for AI
-    addAiMessage({
-      type: 'system',
-      content: 'Transcripción recibida: ' + transcript,
-      isInternal: true
-    });
-    
-    // Trigger AI analysis (this would connect to your AI service)
-    // For now, we'll add a mock response
-    setTimeout(() => {
-      addAiMessage({
-        type: 'ai',
-        content: `Análisis de la transcripción: Se detectan síntomas relacionados con ${getRandomMedicalSuggestion()}. ¿Deseas que profundice en algún aspecto específico?`,
-        suggestions: [
-          'Analizar síntomas principales',
-          'Sugerir diagnósticos diferenciales',
-          'Revisar antecedentes relevantes'
-        ]
-      });
-    }, 2000);
-  };
-  
-  const getRandomMedicalSuggestion = () => {
-    const suggestions = [
-      'síntomas respiratorios',
-      'manifestaciones cardiovasculares',
-      'signos neurológicos',
-      'síntomas gastrointestinales',
-      'manifestaciones dermatológicas'
-    ];
-    return suggestions[Math.floor(Math.random() * suggestions.length)];
-  };
-  
-  const handleStartRecording = async () => {
-    try {
-      const stream = await navigator.mediaDevices.getUserMedia({ 
-        audio: {
-          echoCancellation: true,
-          noiseSuppression: true,
-          sampleRate: 44100
-        } 
-      });
-      
-      setMicPermission('granted');
-      
-      // Setup audio analysis
-      setupAudioAnalysis(stream);
-      
-      // Setup transcription
-      if (transcriptionService === 'browser') {
-        setupBrowserTranscription();
-      }
-      
-      // Setup recording
-      mediaRecorderRef.current = new MediaRecorder(stream, {
-        mimeType: 'audio/webm'
-      });
-      
-      audioChunksRef.current = [];
-      
-      mediaRecorderRef.current.ondataavailable = (event) => {
-        audioChunksRef.current.push(event.data);
-      };
-      
-      mediaRecorderRef.current.onstop = () => {
-        // Create audio blob for potential server-side transcription
-        new Blob(audioChunksRef.current, { type: 'audio/webm' });
-        // Stop all tracks
-        stream.getTracks().forEach(track => track.stop());
-      };
-      
-      mediaRecorderRef.current.start(1000); // Collect data every second
-      startRecording();
-      setRecordingTime(0);
-      
-      logEvent('recording_started', { 
-        service: transcriptionService,
-        time: new Date().toISOString()
-      });
-      
-    } catch (error) {
-      console.error('Error accessing microphone:', error);
-      
-      // Handle different error types
-      if (error.name === 'NotAllowedError') {
-        setMicPermission('denied');
-        logEvent('microphone_error', { error: 'Permission denied by user' });
-      } else if (error.name === 'NotFoundError') {
-        setMicPermission('denied');
-        logEvent('microphone_error', { error: 'No microphone device found' });
-      } else {
-        // For other errors, allow retry
-        setMicPermission('prompt');
-        logEvent('microphone_error', { error: error.message });
-      }
-    }
-  };
-  
-  const handleStopRecording = () => {
-    if (mediaRecorderRef.current && isRecording) {
-      mediaRecorderRef.current.stop();
-      
-      if (recognitionRef.current) {
-        recognitionRef.current.stop();
-      }
-      
-      if (audioContextRef.current) {
-        audioContextRef.current.close();
-      }
-      
-      stopRecording(recordingTime);
-      setAudioLevel(0);
-      
-      logEvent('recording_stopped', { 
-        duration: recordingTime,
-        transcript_length: finalTranscript.length
-      });
-    }
-  };
   
   const formatTime = (seconds) => {
     const minutes = Math.floor(seconds / 60);
@@ -611,11 +202,8 @@ const TranscriptionPanel = () => {
                 </span>
               </div>
               <button
-                onClick={async () => {
-                  const diagnostics = await runMicrophoneDiagnostics();
-                  if (diagnostics.accessTest !== 'success') {
-                    setMicPermission('denied');
-                  }
+                onClick={() => {
+                  runMicrophoneDiagnostics();
                 }}
                 className="ml-4 px-3 py-1 bg-blue-600 hover:bg-blue-700 text-white text-xs font-medium rounded transition-colors"
               >
